+++ RR-Backend/src/main/java/com/scor/rr/service/SearchService.java
@@ -32,6 +32,7 @@ import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.persistence.Query;
 import javax.transaction.Transactional;
+import java.math.BigInteger;
 import java.util.*;
 import java.util.function.BiFunction;
 import java.util.stream.Collectors;
@@ -425,7 +426,7 @@ public class SearchService {
 
     private List<FacContractSearchResult> mapFacContract(List<Object[]> resultList) {
         return resultList.stream().map((r) ->
-                new FacContractSearchResult((String) r[0], (Integer) r[1], (String) r[2], (String) r[3], (String) r[4], (String) r[5], (String) r[6], (Long) r[7])
+                new FacContractSearchResult((String) r[0], (Integer) r[1], (String) r[2], (String) r[3], (String) r[4], (String) r[5], (String) r[6], (BigInteger) r[7])
         ).collect(Collectors.toList());
     }
